emitLLVMKernel: avoid relying on isl set variable names
codegenISL would manually set the names of some set dimensions,
expecting them to be preserved across subsequent operations,
but this is not guaranteed.
Convert the corresponding expressions to isl::ast_expr objects instead.

As a nice bonus, the code generation for subscripts and variables
now share the same mechanism, while before one used isl::ast_expr
objects, while the other used an isl::pw_multi_aff.
As a result, a lot of code can be removed.